\name{ccds.19}
\alias{ccds.19}
\docType{data}
\title{CCDS Reference Data (Build hg19)}
\description{A \code{ref.dat} object for the Consensus Coding Sequence (CCDS) data from UCSC human (\emph{Homo sapiens}) build hg19}
\usage{data(ccds.19)}
\format{A data frame \code{ref.dat} object with 18305 genes on 34 variables (1 Gene column, 1 chromosome column, and 32 sequence context motifs}
\details{Each sequence context motif column corresponds to the enumeration of CCDS bases in a given gene that satifies that motif.  The gene column corresponds to HUGO gene ids.}
\source{\url{http://genome.ucsc.edu/}}
\references{\url{http://www.ncbi.nlm.nih.gov/CCDS/CcdsBrowse.cgi}}
\keyword{datasets}
